refactor: changed the structure of the directory
