+++ components/IngredientCard/CountingPanel.tsx
